move MessageCounter to Message namespace
