+++ src/Commit/BatchCommitter.php
@@ -3,7 +3,7 @@
 namespace Junges\Kafka\Commit;
 
 use Junges\Kafka\Commit\Contracts\Committer;
-use Junges\Kafka\MessageCounter;
+use Junges\Kafka\Message\MessageCounter;
 use RdKafka\Message;
 
 class BatchCommitter implements Committer
