+++ src/Commit/DefaultCommitterFactory.php
@@ -5,7 +5,7 @@ namespace Junges\Kafka\Commit;
 use Junges\Kafka\Commit\Contracts\Committer;
 use Junges\Kafka\Commit\Contracts\CommitterFactory;
 use Junges\Kafka\Config\Config;
-use Junges\Kafka\MessageCounter;
+use Junges\Kafka\Message\MessageCounter;
 use RdKafka\KafkaConsumer;
 
 class DefaultCommitterFactory implements CommitterFactory
