+++ tests/Commit/BatchCommitterTest.php
@@ -4,7 +4,7 @@ namespace Junges\Kafka\Tests\Commit;
 
 use Junges\Kafka\Commit\BatchCommitter;
 use Junges\Kafka\Commit\Contracts\Committer;
-use Junges\Kafka\MessageCounter;
+use Junges\Kafka\Message\MessageCounter;
 use Junges\Kafka\Tests\LaravelKafkaTestCase;
 use RdKafka\Message;
